+++ packages/ui/core-components/src/lib/unsorted/viz/core/_Value.svelte
@@ -36,6 +36,21 @@
 	let selected_value;
 	let error;
 
+	// Value Styling Props:
+	export let color = undefined;
+	let fontColor = '';
+	// Negative value font color:
+	export let redNegatives = false;
+	$: redNegatives = redNegatives === 'true' || redNegatives === true;
+
+	$: if (redNegatives || color) {
+		if (redNegatives && selected_value < 0) {
+			fontColor = 'rgb(220 38 38)';
+		} else if (color) {
+			fontColor = color;
+		}
+	}
+
 	let columnSummary;
 	$: {
 		try {
@@ -106,12 +121,12 @@
 {:else if !error}
 	{#if link}
 		<a class="markdown" href={link}>
-			<span>
+			<span style="color: {fontColor}">
 				{formatValue(selected_value, format_object)}
 			</span>
 		</a>
 	{:else}
-		<span>
+		<span style="color: {fontColor}">
 			{formatValue(selected_value, format_object)}
 		</span>
 	{/if}