+++ packages/datasources/mssql/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @evidence-dev/mssql
 
+## 1.0.9
+
+### Patch Changes
+
+- 16b69f04e: Update mssql to v11 to resolve security vulnerability in @azure/identity
+
 ## 1.0.8
 
 ### Patch Changes