+++ sites/docs/CHANGELOG.md
@@ -1,5 +1,43 @@
 # evidence-docs
 
+## 0.0.20
+
+### Patch Changes
+
+- Updated dependencies [3698967f1]
+  - @evidence-dev/core-components@4.7.2
+  - @evidence-dev/evidence@39.1.1
+
+## 0.0.19
+
+### Patch Changes
+
+- Updated dependencies [3571b0b3a]
+  - @evidence-dev/sdk@1.3.6
+  - @evidence-dev/evidence@39.1.1
+  - @evidence-dev/component-utilities@3.2.6
+  - @evidence-dev/core-components@4.7.1
+
+## 0.0.18
+
+### Patch Changes
+
+- Updated dependencies [08818477d]
+- Updated dependencies [08818477d]
+- Updated dependencies [0558bae48]
+- Updated dependencies [db8326efa]
+- Updated dependencies [731d0bb06]
+- Updated dependencies [db8326efa]
+- Updated dependencies [f5decad6b]
+- Updated dependencies [804949397]
+- Updated dependencies [ee8f135ac]
+- Updated dependencies [3b2e06a0b]
+- Updated dependencies [7cb558c32]
+  - @evidence-dev/core-components@4.7.0
+  - @evidence-dev/sdk@1.3.5
+  - @evidence-dev/evidence@39.1.0
+  - @evidence-dev/component-utilities@3.2.5
+
 ## 0.0.17
 
 ### Patch Changes